+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PPTX"
 uuid = "14a86994-10a4-4a7d-b9ad-ef6f3b1fac6a"
 authors = ["Xander de Vries", "Matthijs Cox"]
-version = "0.11.0"
+version = "0.10.5"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
